requirements_full.txt
# Core Django dependencies
Django==5.2.5
django-cors-headers==4.7.0
django-ninja==1.4.3
djangorestframework==3.16.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Essential packages
requests==2.32.5
pillow==11.3.0
python-dotenv==1.1.1
cryptography==45.0.6
pycryptodome==3.23.0

# AWS and storage
boto3==1.35.36
botocore==1.35.36
django-storages==1.14.2

# Basic ML packages - using binary wheels
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.4.2

# TensorFlow CPU-only
tensorflow-cpu==2.17.0
keras==2.17.0

# OpenCV headless
opencv-python-headless==4.8.1.78

# Other essentials
urllib3==2.5.0
certifi==2025.8.3
python-dateutil==2.9.0.post0
pytz==2025.2
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2